Code clean up in FtpDataStream. Added command channel locking to FtpDataStream. Added MODE_B and MODE_S ftp capabilties. They are set by default, MODE_B gets removed if the MODE B command fails and stream mode is used from then out.
In the future FtpDataChannel will be dropped for FtpDataStream. Will leave both for a while, just know that FtpDataChannel is now legacy code and will not be updated with new features. I will be removing all internal usage of FtpDataChannel in the library and replacing it with FtpDataStream as soon as it's feature complete (it's close).

--HG--
extra : convert_revision : svn%3A7523448a-a8a0-4aac-83b0-0615d70583c7%408457