I've renamed the FtpCommandChannel class to FtpControlConnection to follow terminology used in the RFCs. I realize this is probably going to break some peoples code. It is something I have been hesitant to do but I've decided to move forward with it, no sense in beating around the bush right? I apologize for any inconveniences this causes. Essentially, I did a find and replace on CommandChannel -> ControlConnection. There are several methods and properties which were renamed in the process.
This is the last major change I see for a while. I'm about ready to tag this project as beta, code seems stable enough to do so and I think the API (including class/property/method names!) is stable enough to do so. In the mean time I've got to update documentation on the site to reflect the current state of affairs.

--HG--
extra : convert_revision : svn%3A7523448a-a8a0-4aac-83b0-0615d70583c7%408590