Added new check to GetListing() that looks for modified file times in the future. If it finds one and the MDTM command is supported then the modified date is reloaded using MDTM. This process only happens if you pass the Modify flag list option to GetListing
This check happens because a bug was reported several days back where modified times were in the future. The cause of the problem is that LIST formats omit the year in modified times for dates in the last 6 or so months. As a result, dates from, for example, Nov. 2012 and Dec. 2012, are parsed as Nov. 2013 or Dec. 2013 which is wrong. One possible approach discussed was to just subtract 1 year from the date time object however I find that querying the server using MDTM is the most reliable approach to getting an accurate and correct date time value for the modified date. This comes with an implication in that once the window were the year is omitted passes, the date/time parsed will no longer be in the future and as a result won't be as accurate of a representation as what MDTM gave. This can probably be avoided to some degree by rounding off to the nearest minute.