Check dirPat is not empty string
If `localPath` is just a filename, say 'file.txt', then the directory part returned by `Path.GetDirectoryName(localPath)` will be the empty string. Passing that to `Directory.Exists(dirPath)` will cause an exception.

This change ensures that we never pass the empty string to `Directory.Exists(dirPath)`. We can safely make this change, because if `dirPath` is the empty string, then we are downloading to the current directory, which definitely exists, so there is no need to create a new directory.