Fix file upload EOS bug.
Per the .NET Stream API, Read returns zero ONLY on end-of-stream.

Thus, the file upload should not continue after the inner read/write
loop terminates without error.

Stream.Length should not be relied upon for this purpose, as many
implementations of Stream may be buffered, and there is no guarantee
that the length of the Stream is the size of the file.